+++ dbt_dry_run/node_runner/snapshot_runner.py
@@ -1,9 +1,11 @@
-from typing import Set
+from typing import Callable, List, Optional, Set
+
+from pydantic import BaseModel
 
 from dbt_dry_run.exception import SnapshotConfigException, UpstreamFailedException
 from dbt_dry_run.models import BigQueryFieldMode, BigQueryFieldType, Table, TableField
 from dbt_dry_run.models.dry_run_result import DryRunResult
-from dbt_dry_run.models.manifest import Node
+from dbt_dry_run.models.manifest import Node, NodeConfig
 from dbt_dry_run.models.report import DryRunStatus
 from dbt_dry_run.node_runner import NodeRunner
 from dbt_dry_run.sql.statements import SQLPreprocessor, insert_dependant_sql_literals
@@ -15,32 +17,61 @@ def _check_cols_missing(node: Node, table: Table) -> Set[str]:
     return set(filter(lambda col: col not in table.field_names, node.config.check_cols))
 
 
+class SnapshotField(BaseModel):
+    table_field: TableField
+    filter: Optional[Callable[[NodeConfig], bool]] = None
+
+
 DBT_SNAPSHOT_FIELDS = [
-    TableField(
-        name="dbt_scd_id",
-        type=BigQueryFieldType.STRING,
-        mode=BigQueryFieldMode.NULLABLE,
+    SnapshotField(
+        table_field=TableField(
+            name="dbt_scd_id",
+            type=BigQueryFieldType.STRING,
+            mode=BigQueryFieldMode.NULLABLE,
+        )
     ),
-    TableField(
-        name="dbt_updated_at",
-        type=BigQueryFieldType.TIMESTAMP,
-        mode=BigQueryFieldMode.NULLABLE,
+    SnapshotField(
+        table_field=TableField(
+            name="dbt_updated_at",
+            type=BigQueryFieldType.TIMESTAMP,
+            mode=BigQueryFieldMode.NULLABLE,
+        )
     ),
-    TableField(
-        name="dbt_valid_from",
-        type=BigQueryFieldType.TIMESTAMP,
-        mode=BigQueryFieldMode.NULLABLE,
+    SnapshotField(
+        table_field=TableField(
+            name="dbt_valid_from",
+            type=BigQueryFieldType.TIMESTAMP,
+            mode=BigQueryFieldMode.NULLABLE,
+        )
     ),
-    TableField(
-        name="dbt_valid_to",
-        type=BigQueryFieldType.TIMESTAMP,
-        mode=BigQueryFieldMode.NULLABLE,
+    SnapshotField(
+        table_field=TableField(
+            name="dbt_valid_to",
+            type=BigQueryFieldType.TIMESTAMP,
+            mode=BigQueryFieldMode.NULLABLE,
+        )
+    ),
+    SnapshotField(
+        table_field=TableField(
+            name="is_deleted",
+            type=BigQueryFieldType.STRING,
+            mode=BigQueryFieldMode.NULLABLE,
+        ),
+        filter=lambda config: config.hard_deletes == "new_record",
     ),
 ]
 
 
 class SnapshotRunner(NodeRunner):
     preprocessor = SQLPreprocessor([insert_dependant_sql_literals])
+
+    @staticmethod
+    def _get_snapshot_fields(config: NodeConfig) -> List[TableField]:
+        fields = []
+        for snapshot_field in DBT_SNAPSHOT_FIELDS:
+            if snapshot_field.filter is None or snapshot_field.filter(config):
+                fields.append(snapshot_field.table_field)
+        return fields
 
     @staticmethod
     def _validate_snapshot_config(node: Node, result: DryRunResult) -> DryRunResult:
@@ -100,6 +131,9 @@ class SnapshotRunner(NodeRunner):
         ) = self._sql_runner.query(run_sql)
         result = DryRunResult(node, predicted_table, status, exception)
         if result.status == DryRunStatus.SUCCESS and result.table:
-            result.table.fields = [*result.table.fields, *DBT_SNAPSHOT_FIELDS]
+            result.table.fields = [
+                *result.table.fields,
+                *self._get_snapshot_fields(node.config),
+            ]
             result = self._validate_snapshot_config(node, result)
         return result