+++ integration/projects/test_project_with_snapshots/test_project_with_snapshots.py
@@ -1,8 +1,12 @@
-from dbt_dry_run.node_runner.snapshot_runner import DBT_SNAPSHOT_FIELDS
 from integration.conftest import CompletedDryRun
 from integration.utils import get_report_node_by_id, assert_report_node_has_columns
 
-DBT_SNAPSHOT_COLUMN_NAMES = set([field.name for field in DBT_SNAPSHOT_FIELDS])
+DBT_ALWAYS_ON_SNAPSHOT_COLUMN_NAMES = {
+    "dbt_scd_id",
+    "dbt_updated_at",
+    "dbt_valid_from",
+    "dbt_valid_to",
+}
 
 
 def test_case_invalid_unique_key_fails(dry_run_result: CompletedDryRun):
@@ -20,7 +24,7 @@ def test_case_timestamp_column_succeeds(dry_run_result: CompletedDryRun):
         "snapshot.test_project_with_snapshots.case_timestamp_column_snapshot",
     )
     assert_report_node_has_columns(
-        node, {"a", "b", "c", "updated_at", *DBT_SNAPSHOT_COLUMN_NAMES}
+        node, {"a", "b", "c", "updated_at", *DBT_ALWAYS_ON_SNAPSHOT_COLUMN_NAMES}
     )
 
 
@@ -38,7 +42,9 @@ def test_case_all_succeeds(dry_run_result: CompletedDryRun):
         dry_run_result.report,
         "snapshot.test_project_with_snapshots.case_check_all_snapshot",
     )
-    assert_report_node_has_columns(node, {"a", "b", "c", *DBT_SNAPSHOT_COLUMN_NAMES})
+    assert_report_node_has_columns(
+        node, {"a", "b", "c", *DBT_ALWAYS_ON_SNAPSHOT_COLUMN_NAMES}
+    )
 
 
 def test_case_single_column_succeeds(dry_run_result: CompletedDryRun):
@@ -46,7 +52,9 @@ def test_case_single_column_succeeds(dry_run_result: CompletedDryRun):
         dry_run_result.report,
         "snapshot.test_project_with_snapshots.case_check_single_column_snapshot",
     )
-    assert_report_node_has_columns(node, {"a", "b", "c", *DBT_SNAPSHOT_COLUMN_NAMES})
+    assert_report_node_has_columns(
+        node, {"a", "b", "c", *DBT_ALWAYS_ON_SNAPSHOT_COLUMN_NAMES}
+    )
 
 
 def test_case_invalid_column_fails(dry_run_result: CompletedDryRun):
@@ -56,3 +64,13 @@ def test_case_invalid_column_fails(dry_run_result: CompletedDryRun):
     )
     assert not node.success
     assert node.error_message == "SnapshotConfigException"
+
+
+def test_hard_deletes_creates_is_deleted_column(dry_run_result: CompletedDryRun):
+    node = get_report_node_by_id(
+        dry_run_result.report,
+        "snapshot.test_project_with_snapshots.case_check_hard_deletes",
+    )
+    assert_report_node_has_columns(
+        node, {"a", "b", "c", *DBT_ALWAYS_ON_SNAPSHOT_COLUMN_NAMES, "is_deleted"}
+    )